timeseries: add tests for #9012 and move test file to TS
